Refactor implementation for the chart
