+++ views/nutrition.gohtml
@@ -14,12 +14,13 @@
 <body class="bg-slate-50">
 <main class="w-[450px] mx-auto">
     <h1 class="font-semibold text-4xl text-center my-8">Nutrition</h1>
-    <section data-nutrition="{{ .NutritionJSON }}" id="nutrition-diagram-section" class="bg-white p-5 mb-4 rounded-xl shadow-md">
+    <section id="nutrition-diagram-section" class="bg-white p-5 mb-4 rounded-xl shadow-md">
         <h2 class="font-medium text-xl text-slate-700 mb-2.5">
             Trend
         </h2>
         <canvas
                 id="nutrition-diagram"
+                data-nutrition="{{ .NutritionJSON }}"
                 width="400"
                 height="300"
                 aria-label="Nutrition Diagram"
